+++ crates/hotpath/src/lib_on/debug/gauge.rs
@@ -21,7 +21,7 @@ fn get_thread_id() -> Option<u64> {
 
 #[derive(Debug, Clone)]
 pub struct GaugeEntry {
-    pub id: u64,
+    pub id: u32,
     pub key: String,
     pub source: &'static str,
     pub current_value: f64,
@@ -41,7 +41,7 @@ pub struct GaugeLog {
 
 #[cfg_attr(feature = "hotpath-meta", hotpath_meta::measure_all)]
 impl GaugeEntry {
-    pub fn new(id: u64, key: String, source: &'static str, initial_value: f64) -> Self {
+    pub fn new(id: u32, key: String, source: &'static str, initial_value: f64) -> Self {
         Self {
             id,
             key,
@@ -56,7 +56,7 @@ impl GaugeEntry {
 }
 
 pub struct GaugeHandle {
-    id: u64,
+    id: u32,
     key: String,
     source: &'static str,
 }
@@ -129,7 +129,7 @@ pub fn get_debug_gauge_entries_json() -> Vec<JsonDebugEntry> {
 }
 
 #[cfg_attr(feature = "hotpath-meta", hotpath_meta::measure(log = true))]
-pub fn get_debug_gauge_logs(id: u64) -> Option<JsonDebugGaugeLogs> {
+pub fn get_debug_gauge_logs(id: u32) -> Option<JsonDebugGaugeLogs> {
     let current_elapsed_ns = START_TIME
         .get()
         .map(|t| t.elapsed().as_nanos() as u64)
