+++ crates/hotpath/src/lib_on/functions/alloc/state.rs
@@ -131,7 +131,7 @@ type LogEntry = (
 
 #[derive(Debug, Clone)]
 pub struct FunctionStats {
-    pub id: u64,
+    pub id: u32,
     pub name: &'static str,
     pub count: u64,
     bytes_total_hist: Option<Histogram<u64>>,
@@ -156,7 +156,7 @@ impl FunctionStats {
 
     #[allow(clippy::too_many_arguments)]
     pub fn new_alloc(
-        id: u64,
+        id: u32,
         name: &'static str,
         bytes_total: u64,
         count_total: u64,
@@ -353,7 +353,7 @@ impl FunctionStats {
 pub(crate) struct FunctionsState {
     pub sender: Option<Sender<Measurement>>,
     pub shutdown_tx: Option<Sender<()>>,
-    pub completion_rx: Option<Mutex<Receiver<HashMap<u64, FunctionStats>>>>,
+    pub completion_rx: Option<Mutex<Receiver<HashMap<u32, FunctionStats>>>>,
     pub query_tx: Option<Sender<super::super::FunctionsQuery>>,
 
     pub start_time: Instant,
@@ -363,8 +363,8 @@ pub(crate) struct FunctionsState {
 }
 
 pub(crate) fn process_measurement(
-    stats: &mut HashMap<u64, FunctionStats>,
-    name_to_id: &mut HashMap<&'static str, u64>,
+    stats: &mut HashMap<u32, FunctionStats>,
+    name_to_id: &mut HashMap<&'static str, u32>,
     m: Measurement,
     start_time: Instant,
 ) {
